requirements.txt
torch
numpy
pandas
python-dotenv
scikit-learn
python-binance